started to refactor, then started playing around with a pattern
