+++ python/hoe/color_utils.py
@@ -167,7 +167,7 @@ _BRIGHTNESS = [
 MAX_BRIGHTNESS = len(_BRIGHTNESS)
 
 
-def linear_brightness(val, interpolate=False):
+def linear_brightness(val):
     return _BRIGHTNESS[int(val)]
