+++ python/hoe/utils.py
@@ -1,0 +1,5 @@
+import numpy as np
+
+
+def randrange(low, high):
+    return np.random.rand() * (high - low) + low
